% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{cleanOpenInterfaceData}
\alias{cleanOpenInterfaceData}
\title{cleanOpenInterfaceData}
\usage{
cleanOpenInterfaceData(input)
}
\arguments{
\item{input}{user input}
}
\value{
df
}
\description{
cleanOpenInterfaceData
}
