% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{find_time_start_SDCard}
\alias{find_time_start_SDCard}
\title{find_time_start_SDCard}
\usage{
find_time_start_SDCard(input, time)
}
\arguments{
\item{input}{user input}

\item{time}{time from user input}
}
\description{
find_time_start_SDCard
}
